BE - project fix & refactoring

+++ src/main/java/com/zerobase/foodlier/global/profile/facade/ProfileFacade.java
@@ -1,6 +1,7 @@
 package com.zerobase.foodlier.global.profile.facade;
 
 import com.zerobase.foodlier.common.s3.service.S3Service;
+import com.zerobase.foodlier.common.security.provider.dto.MemberAuthDto;
 import com.zerobase.foodlier.module.member.member.domain.model.Member;
 import com.zerobase.foodlier.module.member.member.local.dto.CoordinateResponseDto;
 import com.zerobase.foodlier.module.member.member.local.service.LocalService;
@@ -10,6 +11,8 @@ import com.zerobase.foodlier.module.member.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
+import java.util.Objects;
+
 @Component
 @RequiredArgsConstructor
 public class ProfileFacade {
@@ -18,18 +21,19 @@ public class ProfileFacade {
     private final LocalService localService;
 
     public void deleteProfileUrlAndGetAddressUpdateProfile(
-            String email,
+            MemberAuthDto memberAuthDto,
             MemberPrivateProfileForm form
     ) {
-        Member member = memberService.findByEmail(email);
+        Member member = memberService.findById(memberAuthDto.getId());
         CoordinateResponseDto coordinateResponseDto = localService.getCoordinate(
-                form.getRoadAddress() != null ? form.getRoadAddress() :
+                Objects.nonNull(form.getRoadAddress()) ? form.getRoadAddress() :
                         member.getAddress().getRoadAddress()
         );
-        String imageUrl = form.getProfileImage() != null ?
+        String imageUrl = Objects.nonNull(form.getProfileImage()) ?
                 s3Service.getImageUrl(form.getProfileImage()) :
                 member.getProfileUrl();
-        if (!imageUrl.equals(member.getProfileUrl())) {
+        if (!imageUrl.equals(member.getProfileUrl())
+                && Objects.nonNull(member.getProfileUrl())) {
             s3Service.deleteImage(member.getProfileUrl());
         }
         memberService.updatePrivateProfile(
